refactor: split into plugin/ and lua/ files
